+++ lib/puppet/type/cs_clone.rb
@@ -18,13 +18,25 @@ Puppet::Type.newtype(:cs_clone) do
     desc 'The corosync resource primitive to be cloned.'
   end
 
+  newproperty(:group) do
+    desc 'The corosync resource group to be cloned.'
+  end
+
   newproperty(:clone_max) do
     desc "How many copies of the resource to start.
       Defaults to the number of nodes in the cluster."
+
+    newvalues(%r{\d+}, :absent)
+
+    defaultto :absent
   end
   newproperty(:clone_node_max) do
     desc "How many copies of the resource can be started on a single node.
     Defaults to 1."
+
+    newvalues(%r{\d+}, :absent)
+
+    defaultto :absent
   end
 
   newproperty(:notify_clones) do
@@ -32,20 +44,26 @@ Puppet::Type.newtype(:cs_clone) do
       and when the action was successful.
       Allowed values: true, false"
 
-    newvalues(:true, :false)
+    newvalues(:true, :false, :absent)
+
+    defaultto :absent
   end
 
   newproperty(:globally_unique) do
     desc "Does each copy of the clone perform a different function?
       Allowed values: true, false"
 
-    newvalues(:true, :false)
+    newvalues(:true, :false, :absent)
+
+    defaultto :absent
   end
 
   newproperty(:ordered) do
     desc 'Should the copies be started in series (instead of in parallel). Allowed values: true, false'
 
-    newvalues(:true, :false)
+    newvalues(:true, :false, :absent)
+
+    defaultto :absent
   end
 
   newproperty(:interleave) do
@@ -53,7 +71,9 @@ Puppet::Type.newtype(:cs_clone) do
       as soon as their peer instance has (rather than waiting for every instance of the other clone has).
       Allowed values: true, false"
 
-    newvalues(:true, :false)
+    newvalues(:true, :false, :absent)
+
+    defaultto :absent
   end
 
   newparam(:cib) do
@@ -77,11 +97,13 @@ Puppet::Type.newtype(:cs_clone) do
     autos
   end
 
-  autorequire(:cs_primitive) do
-    autos = []
-    autos << unmunge_cs_primitive(should(:primitive)) if should(:primitive)
+  { cs_group: :group, cs_primitive: :primitive }.each do |type, property|
+    autorequire(type) do
+      autos = []
+      autos << unmunge_cs_primitive(should(property)) if should(property)
 
-    autos
+      autos
+    end
   end
 
   def unmunge_cs_primitive(name)
@@ -93,6 +115,9 @@ Puppet::Type.newtype(:cs_clone) do
 
   validate do
     return if self[:ensure] == :absent
-    raise Puppet::Error, 'primitive is mandatory' unless self[:primitive]
+    mandatory_single_properties = [:primitive, :group]
+    has_should = mandatory_single_properties.select { |prop| should(prop) }
+    raise Puppet::Error, "You cannot specify #{has_should.join(' and ')} on this type (only one)" if has_should.length > 1
+    raise Puppet::Error, "You must specify #{mandatory_single_properties.join(' or ')}" if has_should.length != 1
   end
 end